backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
requests==2.32.3
python-dotenv==1.0.1
redis==5.0.7
APScheduler==3.10.4
